Remove rowsets generated before converting historical rowsets (#1048)

+++ be/src/olap/data_dir.h
@@ -44,7 +44,7 @@ public:
     Status init();
 
     const std::string& path() const { return _path; }
-    const int64_t path_hash() const { return _path_hash; }
+    const size_t path_hash() const { return _path_hash; }
     bool is_used() const { return _is_used; }
     void set_is_used(bool is_used) { _is_used = is_used; }
     int32_t cluster_id() const { return _cluster_id; }
@@ -139,7 +139,7 @@ private:
 
 private:
     std::string _path;
-    int64_t _path_hash;
+    size_t _path_hash;
     uint32_t _rand_seed;
 
     std::string _file_system;
